backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.0
pydantic-settings==2.3.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.1
xgboost==2.0.3
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
shap==0.51.0
joblib==1.3.2
celery==5.3.4
redis==5.0.1
requests==2.31.0
aiohttp==3.13.4
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
PyJWT==2.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.0.0
pytz==2023.3